+++ cli/cmd/registry.go
@@ -15,6 +15,8 @@ func SpecRegistryToPlugin(registry specs.Registry) managedplugin.Registry {
 		return managedplugin.RegistryGrpc
 	case specs.RegistryDocker:
 		return managedplugin.RegistryDocker
+	case specs.RegistryCloudQuery:
+		return managedplugin.RegistryCloudQuery
 	default:
 		panic("unknown registry " + registry.String())
 	}