+++ plugins/source/gcp/resources/services/compute/projects_fetch.go
@@ -5,8 +5,8 @@ import (
 
 	compute "cloud.google.com/go/compute/apiv1"
 
+	"github.com/cloudquery/cloudquery/plugins/source/gcp/client"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
-	"github.com/cloudquery/plugins/source/gcp/client"
 
 	pb "cloud.google.com/go/compute/apiv1/computepb"
 )