+++ plugins/source/gcp/resources/services/storage/bucket_policies.go
@@ -3,8 +3,8 @@ package storage
 import (
 	pb "cloud.google.com/go/iam"
 	"github.com/apache/arrow/go/v14/arrow"
+	"github.com/cloudquery/cloudquery/plugins/source/gcp/client"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
-	"github.com/cloudquery/plugins/source/gcp/client"
 )
 
 func BucketPolicies() *schema.Table {